+++ import-service/lib/utils.ts
@@ -2,9 +2,9 @@ export const buildResponse = (statusCode: number, body: unknown) => {
   return {
     statusCode,
     headers: {
-      "Accept-Control-Allow-Credentials": true,
       "Accept-Control-Allow-Origin": "*",
       "Accept-Control-Allow-Headers": "*",
+      "Accept-Control-Allow-Methods": "*",
       "Content-Type": "application/json",
     },
     body: JSON.stringify(body),
